exit 211 for the ls at each steps

+++ submission_files/script_generators/type_1/runscript_generators/D_runEvio2hipo.py
@@ -70,6 +70,10 @@ endif
 
 echo "Directory Content After Background Merging:"
 ls -l
+if ($? != 0) then
+  echo ls failure
+  exit 211
+endif
 
 echo "Removing background file"
 rm $bgFile
